# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_handler.SwaggerError:
    properties:
      code:
        example: VALIDATION_FAILED
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
      requestId:
        example: req-abc123
        type: string
      timestamp:
        example: "2024-02-01T10:30:00Z"
        type: string
    type: object
  internal_handler.SwaggerErrorResponse:
    description: Standard error response format
    properties:
      error:
        $ref: '#/definitions/internal_handler.SwaggerError'
    type: object
...
  internal_handler.SwaggerFavoriteRequest:
    description: Toggle favorite request
    properties:
      isFavorite:
        example: true
        type: boolean
    type: object
  internal_handler.SwaggerFavoriteResponse:
    description: Toggle favorite response
    properties:
      isFavorite:
        example: true
        type: boolean
      success:
        example: true
        type: boolean
    type: object
...
  internal_handler.SwaggerRecipe:
    description: Full recipe with ingredients and steps
    properties:
      cookTime:
        example: 20
        type: integer
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      cuisine:
        example: Italian
        type: string
      description:
        example: Classic Italian pasta dish
        type: string
      difficulty:
        enum:
        - easy
        - medium
        - hard
        example: medium
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      ingredients:
        items:
          $ref: '#/definitions/internal_handler.SwaggerRecipeIngredient'
        type: array
      isFavorite:
        example: false
        type: boolean
      prepTime:
        example: 15
        type: integer
      servings:
        example: 4
        type: integer
      sourceType:
        enum:
        - manual
        - video
        - ai
        - photo
        - webpage
        - image
        example: manual
        type: string
      sourceUrl:
        example: https://youtube.com/watch?v=xyz
        type: string
      steps:
        items:
          $ref: '#/definitions/internal_handler.SwaggerRecipeStep'
        type: array
      syncVersion:
        example: 1
        type: integer
      tags:
        example:
        - italian
        - pasta
        - quick
        items:
          type: string
        type: array
      thumbnailUrl:
        example: https://example.com/thumb.jpg
        type: string
      title:
        example: Spaghetti Carbonara
        type: string
      updatedAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      userId:
        example: 550e8400-e29b-41d4-a716-446655440099
        type: string
    type: object
  internal_handler.SwaggerRecipeIngredient:
    description: Recipe ingredient
    properties:
      category:
        enum:
        - dairy
        - produce
        - proteins
        - bakery
        - pantry
        - spices
        - condiments
        - beverages
        - snacks
        - frozen
        - household
        - other
        example: condiments
        type: string
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440001
        type: string
      isOptional:
        example: false
        type: boolean
      name:
        example: Olive oil
        type: string
      notes:
        example: extra virgin preferred
        type: string
      quantity:
        example: 2
        type: number
      recipeId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      sortOrder:
        example: 0
        type: integer
      unit:
        example: tablespoons
        type: string
      videoTimestamp:
        example: 120
        type: integer
    type: object
  internal_handler.SwaggerRecipeListResponse:
    description: Paginated list of recipes
    properties:
      items:
        items:
          $ref: '#/definitions/internal_handler.SwaggerRecipe'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 25
        type: integer
    type: object
  internal_handler.SwaggerRecipeStep:
    description: Recipe step
    properties:
      createdAt:
        example: "2024-02-01T10:30:00Z"
        type: string
      durationSeconds:
        example: 120
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440002
        type: string
      instruction:
        example: Heat olive oil in a large pan over medium heat
        type: string
      recipeId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      stepNumber:
        example: 1
        type: integer
      technique:
        example: sauteing
        type: string
      temperature:
        example: 350F
        type: string
      videoTimestampEnd:
        example: 180
        type: integer
      videoTimestampStart:
        example: 60
        type: integer
    type: object
...
host: api.dishflow.app
info:
  contact:
    email: support@dishflow.app
    name: DishFlow Support
  description: Recipe management API with AI-powered extraction, pantry tracking,
    and shopping lists
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://dishflow.app/terms
  title: DishFlow API
  version: "1.0"
paths:
...
  /recipes:
    get:
      description: Get paginated list of user's recipes
      parameters:
      - default: 20
        description: Items per page (max 50)
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of recipes
          schema:
            $ref: '#/definitions/internal_handler.SwaggerRecipeListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: List user's recipes
      tags:
      - Recipes
    post:
      consumes:
      - application/json
      description: Create a new recipe with ingredients and steps
      parameters:
      - description: Recipe data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerRecipe'
      produces:
      - application/json
      responses:
        "201":
          description: Recipe created successfully
          schema:
            $ref: '#/definitions/internal_handler.SwaggerRecipe'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new recipe
      tags:
      - Recipes
  /recipes/{recipeID}:
    delete:
      description: Soft delete a recipe (can be recovered)
      parameters:
      - description: Recipe UUID
        in: path
        name: recipeID
        required: true
        type: string
      responses:
        "204":
          description: Recipe deleted successfully
        "400":
          description: Invalid recipe ID
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Recipe not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a recipe
      tags:
      - Recipes
    get:
      description: Get full recipe details including ingredients and steps
      parameters:
      - description: Recipe UUID
        in: path
        name: recipeID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Recipe details
          schema:
            $ref: '#/definitions/internal_handler.SwaggerRecipe'
        "400":
          description: Invalid recipe ID
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Recipe not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Get recipe by ID
      tags:
      - Recipes
    put:
      consumes:
      - application/json
      description: Update an existing recipe's details, ingredients, and steps
      parameters:
      - description: Recipe UUID
        in: path
        name: recipeID
        required: true
        type: string
      - description: Updated recipe data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerRecipe'
      produces:
      - application/json
      responses:
        "200":
          description: Recipe updated successfully
          schema:
            $ref: '#/definitions/internal_handler.SwaggerRecipe'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Recipe not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a recipe
      tags:
      - Recipes
  /recipes/{recipeID}/favorite:
    post:
      consumes:
      - application/json
      description: Mark or unmark a recipe as favorite
      parameters:
      - description: Recipe UUID
        in: path
        name: recipeID
        required: true
        type: string
      - description: Favorite status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler.SwaggerFavoriteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Favorite status updated
          schema:
            $ref: '#/definitions/internal_handler.SwaggerFavoriteResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "403":
          description: Access denied
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
        "404":
          description: Recipe not found
          schema:
            $ref: '#/definitions/internal_handler.SwaggerErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle recipe favorite status
      tags:
      - Recipes
...
securityDefinitions:
  BearerAuth:
    description: 'Enter: Bearer {your_token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...